# docs/api/openapi-full.yaml
openapi: 3.1.0
info:
  title: DataSafeS3 REST Admin API
  version: 1.1.0
  summary: Complete JSON REST API served at /api/v1 (excluding S3 XML)
  description: |
    **Full hand-written OpenAPI 3.1** for all Admin JSON routes registered in internal/api/server.go.

    ## Authentication
    | Scheme | Use |
    |--------|-----|
    | **BearerAPIToken** | Integration tokens ds_* (recommended for automation) |
    | **BearerJWT** | JWT from POST /admin/login or OIDC callback (console sessions) |

    Public endpoints: GET /health, POST /admin/login, OIDC bootstrap, GET /public/share/*.

    ## Not included
    - GET /metrics (Prometheus text)
    - S3 XML API on paths outside /api/v1 (SigV4, port 9000)

    ## Errors
    Standard envelope: JSON object with error field (HTTP 4xx/5xx).

    Interactive docs for **Community subset**: GET /api/v1/docs (Swagger UI).

    Canonical file: [openapi-full.yaml](../../api/openapi-full.yaml)
  contact:
    name: DataSafeS3 / Ilya Trachuk
    url: https://github.com/DirektorBani/DataSafeS3
  license:
    name: Apache 2.0
    identifier: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080/api/v1
    description: Via Caddy (console origin)
  - url: http://localhost:9000/api/v1
    description: storage-server direct
security:
  - BearerAPIToken: []
  - BearerJWT: []
components:
  securitySchemes:
    BearerAPIToken:
      type: http
      scheme: bearer
      bearerFormat: ds_
      description: |
        Integration API token with prefix ds_. Create in console: Access → API tokens.
        Inherits creator's role and bucket access.

    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256 JWT from POST /admin/login, MFA completion, or OIDC callback.
  parameters:
...
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 1000, default: 50 }
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        mfa_required: { type: boolean }
        mfa_token: { type: string }
        mfa_setup_required: { type: boolean }
        scheduled_delete_at: { type: string, format: date-time }
        trashed: { type: boolean }
        trash_id: { type: string }
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
    User:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        email: { type: string }
        role: { type: string, enum: [administrator, operator, user] }
        status: { type: string }
        tenant_id: { type: string }
        team_id: { type: string, description: Primary legacy team assignment }
        mfa_enabled: { type: boolean }
    Team:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        created_at: { type: string, format: date-time }
...
  responses:
...
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
paths:
...
  /activity:
    get:
      operationId: listActivity
      summary: Audit activity log
      tags: [Admin]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      parameters:
          - $ref: '#/components/parameters/Offset'
          - $ref: '#/components/parameters/Limit'
      responses:
        '200': { description: Paginated activity }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /users:
    get:
      operationId: listUsers
      summary: List users
      tags: [Admin]
      security:
        - BearerAPIToken: []
        - BearerJWT: []
      parameters:
          - $ref: '#/components/parameters/Offset'
          - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items: { $ref: '#/components/schemas/User' }
                  total: { type: integer }
                  offset: { type: integer }
                  limit: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      operationId: createUser
      summary: Create user
      tags: [Admin]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      responses:
        '201': { description: User created }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /users/{id}:
    put:
      operationId: updateUser
      summary: Update user
      tags: [Admin]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      parameters:
          - name: id
            in: path
            required: true
            schema: { type: string }
      responses:
        '200': { description: Updated }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      operationId: deleteUser
      summary: Delete user
      tags: [Admin]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      parameters:
          - name: id
            in: path
            required: true
            schema: { type: string }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /teams/{id}:
    get:
      operationId: getTeam
      summary: Get team
      tags: [Admin]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      parameters:
          - name: id
            in: path
            required: true
            schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  team: { $ref: '#/components/schemas/Team' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    put:
      operationId: updateTeam
      summary: Rename team
      tags: [Admin]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      parameters:
          - name: id
            in: path
            required: true
            schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  team: { $ref: '#/components/schemas/Team' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      operationId: deleteTeam
      summary: Delete team
      tags: [Admin]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      parameters:
          - name: id
            in: path
            required: true
            schema: { type: string }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /settings/system:
    get:
      operationId: getSystemConfig
      summary: Get system configuration
      tags: [Admin]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      responses:
        '200': { description: SystemConfig object }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    put:
      operationId: putSystemConfig
      summary: Update system configuration
      tags: [Admin]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /search:
    get:
      operationId: search
      summary: Global search
      tags: [Search]
      security:
        - BearerAPIToken: []
        - BearerJWT: []
      parameters:
          - name: q
            in: query
            required: true
            schema: { type: string }
          - $ref: '#/components/parameters/Offset'
          - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  offset: { type: integer }
                  limit: { type: integer }
                  results: { type: array, items: { type: object } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...